requirements.txt
anyio==3.4.0
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
click==8.0.3
cryptography==36.0.0
fastapi==0.70.0
google-api-core==2.2.2
google-auth==2.3.3
google-cloud-secret-manager==2.8.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.42.0
grpcio-status==1.42.0
h11==0.12.0
idna==3.3
libcst==0.3.23
mypy-extensions==0.4.3
Pillow==8.4.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
PyYAML==6.0
requests==2.26.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing-inspect==0.7.1
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
